Resolve 64-bit warnings and compile problems on Hardy.

+++ trunk/examples/cmdline.c
@@ -181,10 +181,10 @@ cmdline_help(const struct cmdline_opt *opts)
 	opt = &opts[saved_index];
 	if (opt->type != CMDLINE_OPT_EOL) {
 		snprintf(buf, sizeof(buf), "-%-*s --%-*s%s%*s    %s",
-		        maxlen1, opt->short_name ? opt->short_name : "",
-		        maxlen2, opt->long_name ? opt->long_name : "",
+		        (int)maxlen1, opt->short_name ? opt->short_name : "",
+		        (int)maxlen2, opt->long_name ? opt->long_name : "",
 		        (maxlen3 > 0) ? " " : "",
-		        maxlen3, opt->arg_name ? opt->arg_name : "",
+		        (int)maxlen3, opt->arg_name ? opt->arg_name : "",
 		        opt->help);
 		saved_index++;
 		return buf;
